requirements.txt
Django>=4.2
djangorestframework>=3.14.0
graphene-django>=3.1.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.6.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
pandas>=2.1.0
PyMuPDF>=1.23.0
python-pptx>=0.6.21
networkx>=3.2
pydantic>=2.5.0
pytest>=7.4.0
pytest-django>=4.7.0
httpx>=0.25.0
olefile>=0.47
tiktoken>=0.5.0